+++ packages/react/src/FronteggProvider.tsx
@@ -1,14 +1,14 @@
 import React, { FC, ReactNode, useCallback, useEffect, useMemo, useRef } from 'react';
 import { initialize } from '@frontegg/js';
 import { FronteggAppOptions } from '@frontegg/types';
-import { FronteggStoreProvider, useAuthRoutes, CustomComponentRegister } from '@frontegg/react-hooks';
+import { FronteggStoreProvider } from '@frontegg/react-hooks';
 import { BrowserRouter, useHistory, UseHistory } from './routerProxy';
 import { ContextHolder, RedirectOptions, FronteggFrameworks } from '@frontegg/rest-api';
 import { AppHolder } from '@frontegg/js/AppHolder';
-import { useQueryKeeper } from './queryKeeper';
 import { isAuthRoute } from '@frontegg/redux-store';
 import sdkVersion from './sdkVersion';
 import ReactPkg from 'react/package.json';
+import { AlwaysRenderInProvider, HistoryObject } from './AlwaysRenderInProvider';
 
 export type FronteggProviderProps = FronteggAppOptions & {
   appName?: string;
@@ -16,29 +16,14 @@ export type FronteggProviderProps = FronteggAppOptions & {
   children?: ReactNode;
 };
 
-type HistoryObject = {
-  push: (path: string) => void;
-  replace: (path: string) => void;
-};
-
 type ConnectorProps = Omit<FronteggProviderProps, 'history'> & {
   history: HistoryObject;
   isExternalHistory?: boolean;
 };
 
-type QueryKeeperWrapperProps = {
-  history: HistoryObject;
-};
-
 export const ConnectorHistory: FC<Omit<ConnectorProps, 'history'>> = (props) => {
   const history = useHistory();
   return <Connector history={history} {...props} />;
-};
-
-export const QueryKeeperWrapper: FC<QueryKeeperWrapperProps> = ({ history }) => {
-  const { signUpUrl } = useAuthRoutes();
-  useQueryKeeper({ routes: { signUpUrl }, history });
-  return <></>;
 };
 
 export const Connector: FC<ConnectorProps> = ({ history, appName, isExternalHistory = false, ...props }) => {
@@ -100,10 +85,12 @@ export const Connector: FC<ConnectorProps> = ({ history, appName, isExternalHist
     <FronteggStoreProvider
       {...({ ...props, app } as any)}
       alwaysVisibleChildren={
-        <>
-          <CustomComponentRegister app={app} themeOptions={props.themeOptions} />
-          {!isExternalHistory && <QueryKeeperWrapper history={history} />}
-        </>
+        <AlwaysRenderInProvider
+          app={app}
+          themeOptions={props.themeOptions}
+          history={history}
+          isExternalHistory={isExternalHistory}
+        />
       }
     />
   );
